feat(fid): small fixes
